requirements.txt
google-cloud-vision
openpyxl
pandas
streamlit
ipykernel
dotenv
